+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.xx.y] - 2025-MM-DD
+
+### Added
+- Ability to resolve unloaded account cells when applying Merkle updates
+
 ## [2.24.3] - 2025-10-23
 
 ## New